refactor: simplify core package & solve treeshaking [this make imports from core verbose- to be solved later]
