--- packages/core/src/db/+schema.register.ts
@@ -1,6 +1,0 @@
-export * from "./expense.schema";
-export * from "./group.schema";
-export * from "./member.schema";
-export * from "./participant.schema";
-export * from "./preference.schema";
-export * from "./user.schema";
