+++ packages/core/src/lib/ai/__tests__/index.ts
