+++ packages/core/src/lib/ai/index.ts
