+++ packages/core/src/lib/ai/models.ts
