+++ packages/core/src/lib/ai/utils.ts
