+++ packages/core/src/lib/drizzle/schema.ts
@@ -1,0 +1,6 @@
+export * from "../../modules/expense/schema";
+export * from "../../modules/group/schema";
+export * from "../../modules/member/schema";
+export * from "../../modules/participant/schema";
+export * from "../../modules/preference/schema";
+export * from "../../modules/user/schema";
