+++ packages/core/src/lib/utils/index.ts
@@ -1,6 +1,3 @@
 export type Prettify<T> = {
   [K in keyof T]: T[K];
 } & {};
-
-export * from "./neverthrow";
-export * from "./effect";
