+++ packages/core/src/lib/utils/string-similarity.ts
