+++ packages/core/src/modules/group/entity.ts
@@ -1,9 +1,13 @@
-import { eq } from "drizzle-orm";
-import { db } from ".";
-import { groupTable } from "./group.schema";
-import { DatabaseReadError, Transaction } from "./utils";
 import { Effect, pipe } from "effect";
-import { requireValueExists, requireManyElements, TaggedError } from "../utils";
+import { DatabaseReadError, Transaction } from "../../lib/drizzle/utils";
+import {
+  requireManyElements,
+  requireValueExists,
+  TaggedError,
+} from "../../lib/effect";
+import { db } from "../../lib/drizzle";
+import { groupTable } from "./schema";
+import { eq } from "drizzle-orm/sql";
 
 class GroupNotFoundError extends TaggedError("GroupNotFoundError") {}
 class MembersNotFoundError extends TaggedError("MembersNotFoundError") {}
