+++ packages/core/src/modules/index.ts
@@ -1,0 +1,4 @@
+export { expenses } from "./expense/entity";
+export { groups } from "./group/entity";
+export { participants } from "./participant/entity";
+export { users } from "./user/entity";
