+++ packages/core/src/modules/member/schema.ts
@@ -1,9 +1,8 @@
-import { groupTable } from "./group.schema";
-import { DrizzleModelTypes } from "./utils";
-
-import { relations } from "drizzle-orm";
 import { pgTable, primaryKey, text, uuid } from "drizzle-orm/pg-core";
-import { createInsertSchema, createSelectSchema } from "drizzle-valibot";
+import { groupTable } from "../group/schema";
+import { DrizzleModelTypes } from "../../lib/drizzle/utils";
+import { createSelectSchema } from "drizzle-valibot";
+import { relations } from "drizzle-orm/relations";
 
 export const memberTable = pgTable(
   "member",
@@ -28,4 +27,4 @@ export type Member = MemberTypes["Select"];
 export const Member = createSelectSchema(memberTable);
 
 export type MemberInsert = MemberTypes["Insert"];
-export const MemberInsert = createInsertSchema(memberTable);
+export const MemberInsert = createSelectSchema(memberTable);
