+++ packages/core/src/modules/preference/schema.ts
@@ -1,7 +1,3 @@
-import { pgTable, primaryKey, uuid } from "drizzle-orm/pg-core";
-import { DrizzleModelTypes } from "./utils";
-import { createInsertSchema, createSelectSchema } from "drizzle-valibot";
-
 /* 
     NOTE: this is a good idea generated by ai
 
@@ -12,6 +8,10 @@ import { createInsertSchema, createSelectSchema } from "drizzle-valibot";
     for when a user is submitting often. this can time out and fall
     back to the default group id
 */
+
+import { pgTable, primaryKey, uuid } from "drizzle-orm/pg-core";
+import { DrizzleModelTypes } from "../../lib/drizzle/utils";
+import { createInsertSchema, createSelectSchema } from "drizzle-valibot";
 
 export const preferenceTable = pgTable(
   "preference",
