+++ packages/web/src/lib/toast.ts
@@ -1,5 +1,3 @@
-import { unwrapOrThrow } from "@blank/core/utils";
-import { ResultAsync } from "neverthrow";
 import { toast, ToastClassnames } from "sonner";
 
 type WithToastOptions<T> = {
