+++ packages/web/src/server/auth/route.ts
@@ -2,8 +2,8 @@ import { getHeader } from "@tanstack/react-start/server";
 import { subjects } from "@blank/auth/subjects";
 import { notFound, redirect } from "@tanstack/react-router";
 import { createServerFn } from "@tanstack/react-start";
-import { users } from "@blank/core/db";
-import { requireValueExists, TaggedError } from "@blank/core/utils";
+import { users } from "@blank/core/modules";
+import { requireValueExists, TaggedError } from "@blank/core/lib/effect/index";
 import { authenticate, openauth } from "@/server/auth/core";
 import { AuthTokens } from "@/server/utils";
 import { Effect, pipe } from "effect";
